# Conflicts:
#	src/main/java/pickyeater/UI/app/mealplanpage/MealPlanUngeneratedPage.java
#	src/main/java/pickyeater/UI/app/userpage/UserEditModePage.java
#	src/main/java/pickyeater/UI/app/userpage/UserPage.java
#	src/main/java/pickyeater/UI/leftbuttons/MainButton.java
#	src/main/java/pickyeater/basics/groceries/PickyGroceriesGenerator.java
#	src/main/java/pickyeater/executors/ExecutorProvider.java
#	src/main/java/pickyeater/executors/MealPlanViewerExecutor.java

+++ src/main/java/pickyeater/basics/user/PickyDailyProgresses.java
@@ -39,7 +39,7 @@ public class PickyDailyProgresses implements DailyProgresses {
     }
 
     @Override
-    public void addBurnedCalories(int calories) {
+    public void addBurnedCalories(float calories) {
         burnedCalories +=calories;
     }
